break generate image and generate layout into their own files
