+++ lib/generate-image.js
@@ -1,0 +1,22 @@
+var mapnik = require('mapnik');
+var assert = require('assert');
+var emptyPNG = new mapnik.Image(1, 1).encodeSync('png');
+
+module.exports = generateImage;
+
+/**
+ * Generate a PNG image with positioned icons on a sprite.
+ *
+ * @param  {ImgLayout}   layout    An {@link ImgLayout} Object used to generate the image
+ * @param  {Function}    callback  Accepts two arguments, `err` and `image` data
+ */
+function generateImage(layout, callback) {
+    assert(typeof layout === 'object' && typeof callback === 'function');
+    if (!layout.items.length) return callback(null, emptyPNG);
+
+    mapnik.blend(layout.items, {
+        width: layout.width,
+        height: layout.height
+    }, callback);
+}
+
